dev_table: Add the STM32G03xxx/04xxx devices
The information is based on the reference manual RM0444 Rev 3 and
AN2606 Rev 44.

Writing the flash has been tested on STM32G031GxU (32kB flash).

The flash size has been set to the largest possible in the group (64kB)
and hence reading will fail for any of the devices smaller than that
(i.e. the 16kB and 32kB variants)