Allow time for udevd in gpio pin setup
The gpio value node will show up immediately after writing to the export
node, however with restrictive permissions until udevd has applied the
relevant rules. So even if a udev rule is giving an user access, it
might be denied if we access the value node too quickly.

This change introduces a delay on the first access of the pin, if it is
not immediately accessible. This delay is potentially in the middle of a
pin sequence, so it is not an ideal solution.

It must be improved later, for instance by setting up all pins in the
sequence in one first pass before starting changing their values.

https://sourceforge.net/p/stm32flash/tickets/116/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>